refactor swipe code into separate module
