+++ keyball/lib/keyball/keyball.h
@@ -17,6 +17,8 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 #pragma once
 
+#include "keyball_swipe.h"
+
 //////////////////////////////////////////////////////////////////////////////
 // Configurations
 
@@ -295,9 +297,6 @@ void keyball_oled_render_keyinfo(void);
 /// inactive layers.
 void keyball_oled_render_layerinfo(void);
 
-// show current swipe status
-void keyball_oled_render_swipe_debug(void);
-
 /// show mouse motion config
 void keyball_oled_render_ballsubinfo(void);
 
@@ -348,65 +347,6 @@ uint16_t keyball_get_cpi(void);
 /// to 34 (3500CPI).
 void keyball_set_cpi(uint16_t cpi);
 
-
-// === Swipe hook (KB-level detect -> user-level action) ===
-typedef enum {
-    KB_SWIPE_NONE = 0,
-    KB_SWIPE_UP,
-    KB_SWIPE_DOWN,
-    KB_SWIPE_RIGHT,
-    KB_SWIPE_LEFT,
-} kb_swipe_dir_t;
-
-// ユーザー定義のモードタグ（KBは中身を解釈しない）
-typedef uint8_t kb_swipe_tag_t;
-
-// user → KB：スワイプセッション開始/終了
-void            keyball_swipe_begin(kb_swipe_tag_t mode_tag);
-void            keyball_swipe_end(void);
-
-// user が参照したい状態
-bool            keyball_swipe_is_active(void);          // 押下中？
-kb_swipe_tag_t  keyball_swipe_mode_tag(void);           // begin() で渡されたタグ
-kb_swipe_dir_t  keyball_swipe_direction(void);          // 現在の方向（未実装の間は常に NONE）
-bool            keyball_swipe_fired_since_begin(void);  // セッション開始以降に1回でも発火したか
-bool            keyball_swipe_consume_fired(void);      // ↑を取得して false に戻す
-
-// KB → user：発火イベント（弱シンボル；実装は user 側。未定義なら呼ばない）
-__attribute__((weak)) void keyball_on_swipe_fire(kb_swipe_tag_t mode_tag, kb_swipe_dir_t dir);
-
-// ==== Swipe runtime params ====
-typedef struct {
-    uint16_t step;     // 発火しきい値（counts）
-    uint8_t  deadzone; // デッドゾーン（counts）
-    bool     freeze;   // スワイプ中ポインタ凍結
-} kb_swipe_params_t;
-
-// 取得・設定
-kb_swipe_params_t keyball_swipe_get_params(void);
-void keyball_swipe_set_step(uint16_t v);
-void keyball_swipe_set_deadzone(uint8_t v);
-void keyball_swipe_set_freeze(bool on);
-void keyball_swipe_toggle_freeze(void);
-
-// --- OLED UI mode ---
-typedef enum { KB_OLED_MODE_NORMAL = 0, KB_OLED_MODE_DEBUG = 1 } kb_oled_mode_t;
-
-void            keyball_oled_mode_toggle(void);
-void            keyball_oled_set_mode(kb_oled_mode_t m);
-kb_oled_mode_t  keyball_oled_get_mode(void);
-
-// --- Swipe Debug pages (for OLED) ---
-void    keyball_swipe_dbg_toggle(void);
-void    keyball_swipe_dbg_show(bool on);
-void    keyball_swipe_dbg_next_page(void);
-void    keyball_swipe_dbg_prev_page(void);
-uint8_t keyball_swipe_dbg_get_page(void);
-
-// ==== Swipe params persistence ====
-bool keyball_swipe_cfg_load(void);   // 起動時に呼ぶ: true=読めた, false=初期化
-void keyball_swipe_cfg_save(void);   // 現在の params を保存
-void keyball_swipe_cfg_reset(void);  // 既定値に戻して保存（=工場出荷）
 
 
 // ---- Keyball専用 EEPROM ブロック（VIA不使用前提）----
